chore: RHIDP-3048: Add release name to dynamic-plugins configmap name
The dynamic-plugins configmap currently has a hard-coded name of
'dynamic-plugins' which prevents having two Developer Hub installations
in the same namespace.

This pull request updates the name of the 'dynamic-plugins' configmap
to include the name of the release such as '<release>-dynamic-plugins'
which will alleviate this issue and allow multiple installations of
Developer Hub per namespace using Helm.