fix: adjust startup, liveness and readiness probes settings (#50)
* Adjust startup, liveness and readiness probes settings

Startup probe settings seem to have been added in the upstream Backstage Chart
in [1], but the current settings do not allow the RHDH Chart for the
liveness probe to be triggered sufficiently enough for the app to be
considered live.
This adjust such settings by accounting for the worst case scenario
where the application might take a bit long to start.

This also aligns the probe endpoints with the upstream chart.

[1] https://github.com/backstage/charts/pull/216

* Update Chart version

* Update charts/backstage/values.yaml

Co-authored-by: Patrick Knight <pknight@redhat.com>
Co-authored-by: Gustavo Lira e Silva <guga.java@gmail.com>

* Add comment explaining why there is no 'initialDelaySeconds' on both liveness and readiness probes

Co-authored-by: Nick Boldt <nboldt@redhat.com>

---------

Co-authored-by: Patrick Knight <pknight@redhat.com>
Co-authored-by: Gustavo Lira e Silva <guga.java@gmail.com>
Co-authored-by: Nick Boldt <nboldt@redhat.com>

fix URLs and bump to 2.24.2 in 1.4 branch

Signed-off-by: RHDH Build (rhdh-bot) <rhdh-bot@redhat.com>