Install the Knative and SonataFlow CRDs in CI as prerequisite
The Orchestrator flavor requires installing the
orchestrator-infra chart as a prerequisite, but
the OpenShift Serverless and Serverless Operators
installed by this chart are available only on OCP
(from the Red Hat Catalog).
And we cannot install the orchestrator-infra-chart
directly here using the
charts/orchestrator-infra/ci/upstream-olm-values.yaml
file because `ct install` on the orchestrator-infra chart
would then fail to install it due to mismatching
resource labels.

For the simple testing that we are doing here on
a vanilla K8s cluster, we only need both the
Knative and SonataFlow CRDs.

Additional nOTES:
- Serverless 1.35 corresponds to Knative 1.18
- Serverless Logic 1.35 corresponds to Sonataflow 1.43